% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddCriteria.R
\name{add_criteria}
\alias{add_criteria}
\title{Add Criteria Values}
\usage{
add_criteria(data)
}
\arguments{
\item{data}{Dataframe produced from clean_data()}
}
\value{
Dataframe with added criteria columns
}
\description{
Adds criteria values to data based on the beneficial use codes
}
