% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_Char_Names.R
\name{AWQMS_Char_Names}
\alias{AWQMS_Char_Names}
\title{AWQMS Characteristic Names}
\usage{
AWQMS_Char_Names(parameters)
}
\arguments{
\item{parameters}{A list of parameters to convert.}
}
\value{
A list of relevant AWQMS parameters
}
\description{
Converts input parameters to AWQMS characteristics names
}
\examples{
AWQMS_Char_Names(parameters = c("Temperature", "Bacteria", "TSS"))
}
