% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_Char_Names.R
\name{AWQMS_to_standard}
\alias{AWQMS_to_standard}
\title{Convert AWQMS Characteristic Names to standard language}
\usage{
AWQMS_to_standard(AWQMS_params)
}
\arguments{
\item{AWQMS_params}{A list of parameters to convert.}
}
\value{
A list of relevant parameters converted to standard language
}
\description{
Converts AWQMS Characteristic Names to standard language
}
\examples{
AWQMS_to_standard(parameters = c("Temperature, water", "Dissolved oxygen (DO)", "Phosphate-phosphorus"))
}
