% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanData.R
\name{CleanData}
\alias{CleanData}
\title{Clean Data for Status and Trends Analysis}
\usage{
CleanData(data)
}
\arguments{
\item{data}{The result of GetData(), expects a STORET formatted dataframe.}
}
\value{
A slimmer dataframe checked for quality controls.
}
\description{
Removes unnecessary variables, checks data for quality, and adds "sample_datetime" and "sample_id" columns.
}
\examples{
CleanData(data = result-from-GetData())
}
