% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion_stats.R
\name{excursion_stats}
\alias{excursion_stats}
\title{Calculate excursion statistics for status periods}
\usage{
excursion_stats(df)
}
\arguments{
\item{df}{Dataframe returned from an assessment fucntion in odeqassessment package.
df must include a grouping column called 'status_period' which can be created by running odeqstatusandtrends::status_period().}
}
\value{
Dataframe of station excursion information
}
\description{
Creates a dataframe with excursion statistics by station, parmameter, and status period.
Excursion statistics include number of excursions, percent excursion, and the max, median, and minimum for results
that contributed to an excursion.
}
\examples{
excursion_stats(df=data_assessed)
}
