% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Locations.R
\name{get_stations_AWQMS}
\alias{get_stations_AWQMS}
\title{Get stations in Oregon from DEQ's Ambient Water Quality Monitoring System for status and trends analysis}
\usage{
get_stations_AWQMS(polygon, exclude.tribal.lands = TRUE,
  stations.channel.name = "STATIONS")
}
\arguments{
\item{polygon}{Shapefile of the area to query}

\item{exclude.tribal.lands}{Whether or not to exclude stations located on tribal lands. Defaults to TRUE.}

\item{stations.channel.name}{The name (in quotes) of your ODBC connection to the DEQLEAD-LIMS/Stations SQL repository. Defaults to "STATIONS".}
}
\value{
A list of stations within a given shapefile.
}
\description{
Queries the ODEQ stations database to pull all available stations within a given shapefile.
}
\examples{
get_stations_AWQMS(polygon = "your-shapefile-here", exclude.tribal.lands = TRUE, stations.channel.name = "STATIONS")
}
