% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Locations.R
\name{get_stations_NWIS}
\alias{get_stations_NWIS}
\title{Get stations in Oregon for status and trends analysis}
\usage{
get_stations_NWIS(polygon, exclude.tribal.lands = TRUE)
}
\arguments{
\item{polygon}{Shapefile of the area to query}

\item{exclude.tribal.lands}{Whether or not to exclude stations located on tribal lands. Defaults to TRUE.}
}
\value{
A list of stations within a given shapefile.
}
\description{
Queries the ODEQ stations database to pull all available stations within a given shapefile.
}
\examples{
get_stations_NWIS(polygon = "your-shapefile-here", exclude.tribal.lands = TRUE)
}
