% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Locations.R
\name{GetStations}
\alias{GetStations}
\title{Get stations in Oregon for status and trends analysis}
\usage{
GetStations(polygon, parameters = NULL, exclude.tribal.lands = TRUE,
  stations.channel.name = "STATIONS")
}
\arguments{
\item{polygon}{Shapefile of the area to query}

\item{parameters}{A list of parameters with which to filter the query.}

\item{exclude.tribal.lands}{Whether or not to exclude stations located on tribal lands. Defaults to TRUE.}

\item{stations.channel.name}{The name (in quotes) of your ODBC connection to the DEQLEAD-LIMS/Stations SQL repository. Defaults to "STATIONS".}
}
\value{
A list of stations within a given shapefile.
}
\description{
Queries the ODEQ stations database to pull all available stations within a given shapefile.
}
\examples{
GetStations(polygon = "your-shapefile-here", exclude.tribal.lands = TRUE, stations.channel.name = "STATIONS")
}
