% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_summary.R
\name{parameter_summary_by_station}
\alias{parameter_summary_by_station}
\title{Create parameter summary table for status and trends report}
\usage{
parameter_summary_by_station(status, sea_ken, stations)
}
\arguments{
\item{status}{Result from status_stns}

\item{sea_ken}{Result of Seasonal Kendall Analysis}

\item{stations}{Result of get_stations_AWQMS}
}
\value{
Dataframe of summarized status and trends information
}
\description{
Create parameter summary table for status and trends report
}
\examples{
parameter_summary(status = status_df, seaken = seasonal_kendall_df)
}
