% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_exceedance.R
\name{percent_exceedance}
\alias{percent_exceedance}
\title{Determine percent exceedance for status years}
\usage{
percent_exceedance(data, year_range = NULL, status_period = 4)
}
\arguments{
\item{data}{Dataframe to determine status from.}

\item{year_range}{Years from which to determine status. If null, the year range for data provided is used.}

\item{status_period}{Number of whole years to include in a status analysis.
The range of years in status years will be binned by this number}
}
\value{
Dataframe of stations with sufficient data
}
\description{
Creates a dataframe with stations, the number of years with data within
the status years, and the percent exceedance of the results by parameter.
}
\examples{
percent_exceedance(data = data.frame, year_range = c(start_year, end_year), status_period = 5)
}
