% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bacteria.R
\name{plot_bacteria}
\alias{plot_bacteria}
\title{Create bacteria plots to display status and trend.}
\usage{
plot_bacteria(data, seaKen, station)
}
\arguments{
\item{data}{Dataframe to determine status from. Must have 'excursion' column generated.}

\item{seaKen}{Results of Seasonal Kendall Analysis}

\item{station}{The station to plot}
}
\value{
dataframe of stations with sufficient data
}
\description{
Create bacteria plots to display status and trend.
}
\examples{
plot_pH(data = data.frame, seaKen, station)
}
