% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temperature.R
\name{plot_temperature}
\alias{plot_temperature}
\title{Create temperature plots to display status and trend.}
\usage{
plot_temperature(data, seaKen, station)
}
\arguments{
\item{data}{Dataframe to determine status from. Must have 'exceed' column generated.}

\item{seaKen}{Results of Seasonal Kendall Analysis}
}
\value{
dataframe of stations with sufficient data
}
\description{
Create temperature plots to display status and trend.
}
