% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TP.R
\name{plot_TP}
\alias{plot_TP}
\title{Create Total Phosphorus plots to display status and trend.}
\usage{
plot_TP(data, seaKen, station)
}
\arguments{
\item{data}{Dataframe to determine status from. Must have 'excursion' column generated.}

\item{seaKen}{Results of Seasonal Kendall Analysis}

\item{station}{The station to plot}
}
\value{
dataframe of stations with sufficient data
}
\description{
Create Total Phosphorus plots to display status and trend.
}
