% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sea_ken.R
\name{sea_ken}
\alias{sea_ken}
\title{Run Seasonal Kendall trend analysis for all stations with sufficient data}
\usage{
sea_ken(data)
}
\arguments{
\item{data}{Dataframe to conduct a Seasonal Kendall Trend Analysis.}
}
\value{
dataframe of stations with the results of the Seasonal Kendall Trend Analysis
}
\description{
Run Seasonal Kendall trend analysis for all stations with sufficient data
}
