% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_Char_Names.R
\name{simpleCap}
\alias{simpleCap}
\title{Capitalize the first letter of every word in a string}
\usage{
simpleCap(string)
}
\arguments{
\item{string}{A character string to capitalize}
}
\value{
A capitalized string
}
\description{
Takes a given string and capitalizes the first letter of every word
}
\examples{
simpleCap(string = "total suspended solids")
}
