% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_periods.R
\name{status_periods}
\alias{status_periods}
\title{Define status assessment periods}
\usage{
status_periods(datetime, periods = 4, year_range = NULL,
  bins_only = FALSE)
}
\arguments{
\item{datetime}{Vector of datetimes in POSIXct format.}

\item{periods}{Number of shorter intervals to divide the total assesment period into. Default is 4.}

\item{year_range}{Vector of the minimum and maximum years that define the total assessment period. If NULL, the year range provided in datetime is used.}

\item{bins_only}{If TRUE, will return a vector of the each unique character values that define the calendar year range for each assessment period. Default is FALSE.}
}
\value{
Vector of character values that define the calendar year range for each assessment period
}
\description{
Divides the entire assessment period as defined by year_range into shorter intervals based on the number of periods and
codes the datetime values according to which interval they belong to.
}
\examples{
status_periods(datetime = data_assessed$sample_datetime, periods=4, year_range = c(start_year:end_year))

}
