% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_stns.R
\name{status_stns}
\alias{status_stns}
\title{Determine stations with sufficient data for status analysis}
\usage{
status_stns(df)
}
\arguments{
\item{df}{Dataframe returned from an assessment function in odeqassessment package.
df must include a grouping column called 'status_period' which can be created by running odeqstatusandtrends::status_period().}
}
\value{
Dataframe of stations with sufficient data
}
\description{
Creates a dataframe with stations, the number of years with data within
the status years, and whether or not there were any exceedances.
}
\examples{
status_stns(df = data_assessed)
}
