% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdl_targets.R
\docType{data}
\name{tmdl_targets}
\alias{tmdl_targets}
\title{Oregon TMDL Targets}
\format{A dataframe}
\usage{
data(tmdl_targets)
}
\description{
Total Maximum Daily Load water column targets in Oregon, USA. NOTE: This is an incomplete list.
}
\examples{
data(tmdl_targets)
}
\keyword{datasets}
\keyword{tmdl}
