% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_stns.R
\name{trend_stns}
\alias{trend_stns}
\title{Determine stations with sufficient data for trend analysis}
\usage{
trend_stns(data, trend_years = c(format(min(data$sample_datetime),
  "\%Y"):format(Sys.Date(), "\%Y")))
}
\arguments{
\item{data}{Dataframe to determine trend years from.}

\item{trend.years}{Which years to determine status by. Default is current year minus 2 to present year.}
}
\value{
dataframe of stations with sufficient data
}
\description{
Determine stations with sufficient data for trend analysis
}
