% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion_stats.R
\name{excursion_stats}
\alias{excursion_stats}
\title{Calculate excursion statistics for status years}
\usage{
excursion_stats(data, year_range = NULL)
}
\arguments{
\item{data}{Dataframe of data to calculate excursion stats from.}

\item{year_range}{Years from which to determine status. If null, the year range for data provided is used.
The range of years in status years will be binned by this number}
}
\value{
Dataframe of stations
}
\description{
Creates a dataframe with excursion statistics by station, parmameter, and status year bin.
Excursion statistics include number of excursions, percent excursion, and the max, median, and minimum for results
that contributed to an excursion.
}
\examples{
excursion_stats(data = data.frame, year_range = c(start_year, end_year))
}
