% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_summary_map.R
\name{parameter_summary_map}
\alias{parameter_summary_map}
\title{Create parameter summary map for status and trends report}
\usage{
parameter_summary_map(param_summary, area)
}
\arguments{
\item{param_summary}{Parameter summary table from parameter_summary()}

\item{area}{Shapefile for assessment areas}
}
\value{
map of status and trends results
}
\description{
Create parameter summary map for status and trends report
}
\examples{
parameter_summary_map(param_summary = parameter_summary_df)
}
