% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_report.R
\name{st_report}
\alias{st_report}
\title{Create Status and Trends Report}
\usage{
st_report(basin, param_summary, format = "html_document", file_name,
  out_path, complete_years, hucs)
}
\arguments{
\item{basin}{The name of the basin for reporting}

\item{param_summary}{A parameter summary table as returned from odeqstatusandtrends::parameter_summary().}

\item{format}{The output format for the rmarkdown document}

\item{file_name}{The name of the output .html file}

\item{out_path}{The output path}

\item{complete_years}{A list of complete years within analysis}

\item{hucs}{A list of huc8s within analysis area}
}
\value{
An html report located in the output path
}
\description{
Creates a Status and Trends Report with the given parameter summary table, Basin name, and output directory.
}
\examples{
st_report(basin = "Willamette", param_summary = param_sum_Willamette, out_path = ".../your-output-path")
}
