% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_report.R
\name{st_report}
\alias{st_report}
\title{Create Status and Trends Report}
\usage{
st_report(format = "word_document", table_format = "pandoc",
  file_name = "Oregon_SandT_report",
  rmd_dir = "N:/Status_and_Trend_Reports/Report_Files/st_report_files",
  out_dir = "N:/Status_and_Trend_Reports/2019/Statewide Report")
}
\arguments{
\item{format}{The output format for the rmarkdown document}

\item{table_format}{The format for kable tables. "pandoc" is the default. See knitr::kable() for more options}

\item{file_name}{The name of the output .html file}

\item{rmd_dir}{The path to the directory where the status and trend Rmarkdown files are saved.}

\item{out_path}{The output path}
}
\value{
A report located in the out_path
}
\description{
Creates a Status and Trends Report with the given parameter summary table, Basin name, and output directory.
}
\examples{
st_report(format = "word_document", table_format = "pandoc", file_name = "Oregon_SandT_report",
out_path = "N:/Status_and_Trend_Reports/2019/Statewide Report")
}
