% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Locations.R
\name{StationsInPoly}
\alias{StationsInPoly}
\title{Clip stations to a polygon}
\usage{
StationsInPoly(stations, polygon, outside = FALSE)
}
\arguments{
\item{stations}{Dataframe of stations with latitude, longitude, and source datum.}

\item{polygon}{Shapefile of the area to query}

\item{outside}{Set to true if you want stations located outside of the polygon instead. Default is FALSE.}
}
\value{
A list of stations inside, or outside if outside=TRUE, of a given polygon.
}
\description{
Returns the stations located within a given shapefile, or outside if outside=TRUE.
}
\examples{
StationsInPoly(stations = "result-of-GetStations()", polygon = "your-shapefile-here", outside=FALSE)
}
