% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_stns.R
\name{status_stns}
\alias{status_stns}
\title{Determine stations with sufiicient data for status analysis}
\usage{
status_stns(data, status_years = c((as.numeric(format(Sys.Date(), "\%Y"))
  - 2):format(Sys.Date(), "\%Y")))
}
\arguments{
\item{data}{Dataframe to determine status from.}

\item{status.years}{Which years to determine status by. Default is current year minus 2 to present year.}
}
\value{
dataframe of stations with sufficient data
}
\description{
Determine stations with sufiicient data for status analysis
}
