% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_stns.R
\name{status_stns}
\alias{status_stns}
\title{Determine stations with sufficient data for status analysis}
\usage{
status_stns(data, year_range = c(year(Sys.Date()) - 21, year(Sys.Date())
  - 1), status_period = 4)
}
\arguments{
\item{data}{Dataframe to determine status from.}

\item{year_range}{Years from which to determine status. Default is most
recent full year and the 20 years previous.}

\item{status_period}{Number of whole to include in a status analysis.
The range of years in status years will be binned by this number}
}
\value{
Dataframe of stations with sufficient data
}
\description{
Creates a dataframe with stations, the number of years with data within
the status years, and whether or not there were any exceedances.
}
\examples{
status_stns(data = data.frame, year_range = c(start_year, end_year), status_period = 5)
}
