% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_stns.R
\name{trend_stns}
\alias{trend_stns}
\title{Determine stations with sufficient data for trend analysis}
\usage{
trend_stns(df, trend_years = NULL)
}
\arguments{
\item{data}{Dataframe to determine stations for trend analysis}

\item{trend.years}{Which years to determine trend by. Default is the minimum year within the data to the current year.}
}
\value{
Dataframe of stations with sufficient years of data
}
\description{
Creates a dataframe with stations and the number of years with data. Stations with less than 8 different years of data
are not sufficient for trend analysis via the status and trends methods.
}
\examples{
trend_stns(df = data_assessed, trend_years = c(format(min(data$sample_datetime), "\%Y"):format(Sys.Date(), "\%Y")))
}
